requirements.txt
django-filter==0.15.1
django-timezone-field==2.0
Django==1.10.6
djangorestframework==3.4.7
drf-nested-routers==0.11.1
Markdown==2.6.7
Pillow==3.4.2
psycopg2==2.6.2
python-social-auth==0.2.21
pytz==2016.10
